+++ CHANGES.rst
@@ -32,6 +32,7 @@
 - Disabled threads parameter for tSNE (#130)
 - Added 'plot_starsolo_quality' and 'plot_starsolo_UMI' to plotting module (#78)
 - Fixed issues with clustered dotplot with new code (#122)
+- Added "top_genes" parameter to tools.marker_genes.get_rank_genes_tables, and set the default to 200 (#153)
 
 Changes to notebooks
 --------------------